+++ src/Aevatar.HomeAssistantClient/Services/Services_services_see_fields_gps_accuracy_selector_number.cs
@@ -15,7 +15,7 @@ namespace Aevatar.HomeAssistantClient.Services
         /// <summary>Stores additional data not described in the OpenAPI description found when deserializing. Can be used for serialization as well.</summary>
         public IDictionary<string, object> AdditionalData { get; set; }
         /// <summary>The min property</summary>
-        public int? Min { get; set; }
+        public double? Min { get; set; }
         /// <summary>The mode property</summary>
 #if NETSTANDARD2_1_OR_GREATER || NETCOREAPP3_1_OR_GREATER
 #nullable enable
@@ -25,7 +25,7 @@ namespace Aevatar.HomeAssistantClient.Services
         public string Mode { get; set; }
 #endif
         /// <summary>The step property</summary>
-        public int? Step { get; set; }
+        public double? Step { get; set; }
         /// <summary>The unit_of_measurement property</summary>
 #if NETSTANDARD2_1_OR_GREATER || NETCOREAPP3_1_OR_GREATER
 #nullable enable
@@ -59,9 +59,9 @@ namespace Aevatar.HomeAssistantClient.Services
         {
             return new Dictionary<string, Action<IParseNode>>
             {
-                { "min", n => { Min = n.GetIntValue(); } },
+                { "min", n => { Min = n.GetDoubleValue(); } },
                 { "mode", n => { Mode = n.GetStringValue(); } },
-                { "step", n => { Step = n.GetIntValue(); } },
+                { "step", n => { Step = n.GetDoubleValue(); } },
                 { "unit_of_measurement", n => { UnitOfMeasurement = n.GetStringValue(); } },
             };
         }
@@ -72,9 +72,9 @@ namespace Aevatar.HomeAssistantClient.Services
         public virtual void Serialize(ISerializationWriter writer)
         {
             _ = writer ?? throw new ArgumentNullException(nameof(writer));
-            writer.WriteIntValue("min", Min);
+            writer.WriteDoubleValue("min", Min);
             writer.WriteStringValue("mode", Mode);
-            writer.WriteIntValue("step", Step);
+            writer.WriteDoubleValue("step", Step);
             writer.WriteStringValue("unit_of_measurement", UnitOfMeasurement);
             writer.WriteAdditionalData(AdditionalData);
         }
