+++ src/Aevatar.HomeAssistantClient/Services/Services_services_toggle_fields_color_temp_kelvin_selector_color_temp.cs
@@ -15,9 +15,9 @@ namespace Aevatar.HomeAssistantClient.Services
         /// <summary>Stores additional data not described in the OpenAPI description found when deserializing. Can be used for serialization as well.</summary>
         public IDictionary<string, object> AdditionalData { get; set; }
         /// <summary>The max property</summary>
-        public int? Max { get; set; }
+        public double? Max { get; set; }
         /// <summary>The min property</summary>
-        public int? Min { get; set; }
+        public double? Min { get; set; }
         /// <summary>The unit property</summary>
 #if NETSTANDARD2_1_OR_GREATER || NETCOREAPP3_1_OR_GREATER
 #nullable enable
@@ -51,8 +51,8 @@ namespace Aevatar.HomeAssistantClient.Services
         {
             return new Dictionary<string, Action<IParseNode>>
             {
-                { "max", n => { Max = n.GetIntValue(); } },
-                { "min", n => { Min = n.GetIntValue(); } },
+                { "max", n => { Max = n.GetDoubleValue(); } },
+                { "min", n => { Min = n.GetDoubleValue(); } },
                 { "unit", n => { Unit = n.GetStringValue(); } },
             };
         }
@@ -63,8 +63,8 @@ namespace Aevatar.HomeAssistantClient.Services
         public virtual void Serialize(ISerializationWriter writer)
         {
             _ = writer ?? throw new ArgumentNullException(nameof(writer));
-            writer.WriteIntValue("max", Max);
-            writer.WriteIntValue("min", Min);
+            writer.WriteDoubleValue("max", Max);
+            writer.WriteDoubleValue("min", Min);
             writer.WriteStringValue("unit", Unit);
             writer.WriteAdditionalData(AdditionalData);
         }
